Resolve formatting issues, never rely on order of operations, use parenthesis, use 16 bits for auth attempts, 8 bits is not enough for configuration

+++ src/TAO/API/types/session.h
@@ -289,7 +289,7 @@ namespace TAO
 
 
                 /** Number of incorrect authentication attempts recorded for this session **/
-                uint8_t nAuthAttempts;
+                uint16_t nAuthAttempts;
 
 
                 /** Encrypted pointer of signature chain **/
